fix(context): readd wrapper type `ContextRef`
